Fixing a wrong datatype for storing vertices values. Should be float instead of int.
git-svn-id: https://www.underworldproject.org/svn/3DVisualizer@29 bb334f07-60d4-4302-955a-b960dd4c4441

+++ 1.7/Concrete/UnderworldHDF5File.cpp
@@ -340,7 +340,7 @@ Visualization::Abstract::DataSet* UnderworldHDF5File::load(const std::vector<std
    vertNodeCount[0]=1;
    vertNodeCount[1]=vertDims[1];
    hid_t vertMemSpace=H5Screate_simple(vertRank,vertNodeCount,NULL);
-   int* vertBuffer=new int[vertDims[1]];
+   float* vertBuffer=new float[vertDims[1]];
    for(int vert_I=0;vert_I<vertDims[0];++vert_I)
       {
       vertStart[0]=vert_I;
